feat: moving types to own folder
